Some optimizations at bloom filter
*bloom_filter_set_to_zero now is called less times
*bloom_filter_set_to_zero now uses memset to zero out the bit vector

+++ part_2/src/bloomfilter.h
@@ -9,7 +9,7 @@
 
 typedef struct BloomFilter {
     char *bitVector;
-    int bitVectorSize;
+    size_t bitVectorSize;
     int numberOfHashFunctions;
     double expectedProbFalsePositives;
 } BloomFilter;
